backend/requirements.txt
fastapi
uvicorn
openai
chromadb
litellm
python-dotenv
